Rename validators submodule
