+++ src/geff/validate/graph.py
