+++ src/geff/validate/segmentation.py
